backend/requirements.txt
Flask~=3.1.0
dotenv~=0.9.9
python-dotenv~=1.0.1
pymongo~=4.11.2
Werkzeug~=3.1.3
PyJWT~=2.10.1